# Conflicts:
#	viewmodel/src/test/java/com/baghdad/viewmodel/movieDetils/MovieDetailsViewModelTest.kt

+++ remote_datasource/src/main/java/com/baghdad/remoteDataSource/response/ReviewsResponse.kt
@@ -36,5 +36,5 @@ data class MovieAuthorDetails(
     @SerializedName("avatar_path")
     val avatarPath: String? = null,
     @SerializedName("rating")
-    val rating: Float? = null
+    val rating: Double? = null
 )
